split json dep into seperate file excluded with no_json (for grol)
