+++ .gitignore
@@ -1,3 +1,4 @@
+.DS_Store
 .golangci.yml
 # Binaries for programs and plugins
 *.exe
